Updated business licence page and form state
