+++ prime-dotnet-webapi/ViewModels/Health Authority Sites/HealthAuthoritySiteViewModel.cs
@@ -14,7 +14,6 @@ namespace Prime.ViewModels.HealthAuthoritySites
         public DateTimeOffset? SubmittedDate { get; set; }
         public DateTimeOffset? ApprovedDate { get; set; }
         public AddressViewModel PhysicalAddress { get; set; }
-
         public ICollection<BusinessDay> BusinessHours { get; set; }
         // public ICollection<RemoteUser> RemoteUsers { get; set; }
 
@@ -22,7 +21,6 @@ namespace Prime.ViewModels.HealthAuthoritySites
         public HealthAuthorityCareTypeViewModel HealthAuthorityCareType { get; set; }
         public int? HealthAuthorityPharmanetAdministratorId { get; set; }
         public int? HealthAuthorityTechnicalSupportId { get; set; }
-
         public bool Completed { get; set; }
     }
 }
